chromium/third_party/blink/web_tests/virtual/out-of-blink-cspee/external/wpt/content-security-policy/embedded-enforcement/subsumption_algorithm-host_sources-paths-expected.txt

This is a testharness.js-based test.
FAIL Returned CSP must specify a path. assert_equals: expected (undefined) undefined but got (boolean) true
PASS Returned CSP has a more specific path.
PASS Matching paths.
FAIL Empty path is not subsumed by specified paths. assert_equals: expected (undefined) undefined but got (boolean) true
PASS All specific paths match except the order.
PASS Returned CSP allows only one path.
PASS `/` path should be subsumed by an empty path.
PASS Unspecified path should be subsumed by `/`.
FAIL That should not be true when required csp specifies a specific page. assert_unreached: No message should be sent from the frame. Reached unreachable code
Harness: the test ran to completion.
